==> 03_impeachmentanalysis_markdown.Rmd <==
---
title: "Impeachment Analysis"
author: "AK"
date: ""
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(janitor)
library(summarytools)
library(glue)
options(scipen = 999)

working_joined <- readRDS("output/joined_impeachment.rds")


```

Analysis of bill sponsorship vs. not


Trump- or Hillary-won districts in 2016 (R or D)

```{r, echo=FALSE}

print(ctable(working_joined$p16winningparty,
             working_joined$for_impeachment, 
             prop = "c", headings = FALSE), method = "render")



```


Seats won by close margins in 2018

```{r, echo=FALSE}

print(ctable(working_joined$margin_flag,
             working_joined$for_impeachment, 
             prop = "r", headings = FALSE), method = "render")


```





GDP in district vs. national average

```{r, echo=FALSE}

print(ctable(working_joined$gdp_above_national,
             working_joined$for_impeachment, 
             prop = "c", headings = FALSE), method = "render")


```


Percentage with bachelor's degree or higher vs. national average

```{r, echo=FALSE}

print(ctable(working_joined$pct_bachelors_compared_to_national,
             working_joined$for_impeachment, 
             prop = "c", headings = FALSE), method = "render")


```


Percentage of non-white population vs. national average

```{r, echo=FALSE}

print(ctable(working_joined$pct_nonwhite_compared_to_national,
             working_joined$for_impeachment, 
             prop = "c", headings = FALSE), method = "render")


```


Is more than 20% of the district's area considered rural 

```{r, echo=FALSE}

print(ctable(working_joined$rural_pop_above20pct,
             working_joined$for_impeachment, 
             prop = "c", headings = FALSE), method = "render")

```







<!-- ## Dems from Trump Districts only -->

<!-- ```{r} -->
<!-- trump_dists_only <- working_joined %>%  -->
<!--   filter(p16winningparty == "R") -->

<!-- ``` -->

<!-- ```{r} -->

<!-- print(ctable(trump_dists_only$medincome.abovebelow.natl, -->
<!--              trump_dists_only$stance,  -->
<!--              prop = "r", headings = FALSE), method = "render") -->


<!-- ``` -->


<!-- ```{r} -->

<!-- print(ctable(trump_dists_only$pct.ed.college.all.abovebelow.natl, -->
<!--              trump_dists_only$stance,  -->
<!--              prop = "r", headings = FALSE), method = "render") -->


<!-- ``` -->

<!-- ```{r} -->

<!-- print(ctable(trump_dists_only$margin_flag, -->
<!--              trump_dists_only$stance,  -->
<!--              prop = "r", headings = FALSE), method = "render") -->


<!-- ``` -->


<!-- #### Dems on the bill from trump districts -->

<!-- ```{r} -->

<!-- trump_dists_onbill <- working_joined %>%  -->
<!--   filter(p16winningparty == "R", -->
<!--          stance == "sponsoring") -->

<!-- # view(dfSummary(trump_dists_onbill)) -->


<!-- ``` -->



<!-- #### Dems OFF the bill from trump districts -->

<!-- ```{r} -->

<!-- trump_dists_offbill <- working_joined %>%  -->
<!--   filter(p16winningparty == "R", -->
<!--          stance == "not_sponsoring") -->

<!-- # view(dfSummary(trump_dists_offbill)) -->


<!-- ``` -->
